move molecular_hamiltona to hamiltonian
