+++ pennylane/qchem/__init__.py
@@ -21,14 +21,18 @@ from .openfermion_obs import (
     dipole_of,
     meanfield,
     decompose,
-    molecular_hamiltonian,
 )
 from .basis_data import load_basisset
 from .basis_set import BasisFunction, atom_basis_data, mol_basis_data
 from .convert import import_operator, import_state
 from .dipole import dipole_integrals, fermionic_dipole, dipole_moment
 from .factorization import basis_rotation, factorize
-from .hamiltonian import electron_integrals, fermionic_hamiltonian, diff_hamiltonian
+from .hamiltonian import (
+    electron_integrals,
+    fermionic_hamiltonian,
+    diff_hamiltonian,
+    molecular_hamiltonian,
+)
 from .hartree_fock import scf, nuclear_energy, hf_energy
 from .integrals import (
     primitive_norm,
